+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 2.3.0 - 2017-2-22
 
+- Update `bounds` Typescript definition
 - Drop `extent` input, now `bounds` can handle BBox|BBox[]|GeoJSON
 
 ## 2.2.0 - 2017-2-20